emotion_service/requirements.txt
flask>=2.0.0
requests
python-dotenv
gunicorn
